chore: Move tiles to their respective section
