+++ lib/screens/settings/desktop/application.dart
@@ -19,11 +19,13 @@
 
 import 'package:bluecherry_client/providers/settings_provider.dart';
 import 'package:bluecherry_client/screens/settings/desktop/settings.dart';
-import 'package:bluecherry_client/screens/settings/shared/date_language.dart';
 import 'package:bluecherry_client/screens/settings/shared/options_chooser_tile.dart';
+import 'package:bluecherry_client/utils/extensions.dart';
 import 'package:bluecherry_client/widgets/misc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:flutter_localized_locales/flutter_localized_locales.dart';
+import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
 
 class ApplicationSettings extends StatelessWidget {
@@ -135,5 +137,128 @@ class ApplicationSettings extends StatelessWidget {
         ),
       ),
     ]);
+  }
+}
+
+class LanguageSection extends StatelessWidget {
+  const LanguageSection({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    final loc = AppLocalizations.of(context);
+    final settings = context.watch<SettingsProvider>();
+    final currentLocale = Localizations.localeOf(context);
+    const locales = AppLocalizations.supportedLocales;
+    final names = LocaleNames.of(context)!;
+
+    return DropdownButtonHideUnderline(
+      child: ListTile(
+        contentPadding: DesktopSettings.horizontalPadding,
+        leading: CircleAvatar(
+          backgroundColor: Colors.transparent,
+          foregroundColor: theme.iconTheme.color,
+          child: const Icon(Icons.language),
+        ),
+        title: Text(loc.language),
+        trailing: DropdownButton<Locale>(
+          value: currentLocale,
+          onChanged: (value) => settings.locale = value!,
+          items: locales.map((locale) {
+            final name =
+                names.nameOf(locale.toLanguageTag()) ?? locale.toLanguageTag();
+            final nativeName = LocaleNamesLocalizationsDelegate
+                    .nativeLocaleNames[locale.toLanguageTag()] ??
+                locale.toLanguageTag();
+            return DropdownMenuItem<Locale>(
+              value: locale,
+              child: Padding(
+                padding: const EdgeInsetsDirectional.only(end: 8.0),
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    Text(
+                      name.uppercaseFirst,
+                      maxLines: 1,
+                      softWrap: false,
+                      style: theme.textTheme.bodyMedium,
+                    ),
+                    Text(
+                      nativeName.uppercaseFirst,
+                      style: theme.textTheme.labelSmall,
+                    ),
+                  ],
+                ),
+              ),
+            );
+          }).toList(),
+        ),
+      ),
+    );
+  }
+}
+
+class DateFormatSection extends StatelessWidget {
+  const DateFormatSection({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final settings = context.watch<SettingsProvider>();
+    final locale = Localizations.localeOf(context).toLanguageTag();
+    final formats = [
+      'dd MMMM yyyy',
+      'EEEE, dd MMMM yyyy',
+      'EE, dd MMMM yyyy',
+      'MM/dd/yyyy',
+      'dd/MM/yyyy',
+      'MM-dd-yyyy',
+      'dd-MM-yyyy',
+      'yyyy-MM-dd'
+    ].map((e) => DateFormat(e, locale));
+
+    return OptionsChooserTile(
+      title: 'Date Format',
+      description: 'What format to use for displaying dates',
+      icon: Icons.calendar_month,
+      value: '',
+      values: formats.map((format) {
+        return Option(
+          value: format.pattern,
+          text: format.format(DateTime.utc(1969, 7, 20, 14, 18, 04)),
+        );
+      }),
+      onChanged: (v) {
+        settings.dateFormat = DateFormat(v!, locale);
+      },
+    );
+  }
+}
+
+class TimeFormatSection extends StatelessWidget {
+  const TimeFormatSection({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final settings = context.watch<SettingsProvider>();
+    final locale = Localizations.localeOf(context).toLanguageTag();
+
+    final patterns = ['HH:mm', 'hh:mm a'].map((e) => DateFormat(e, locale));
+    final date = DateTime.utc(1969, 7, 20, 14, 18, 04);
+    return OptionsChooserTile(
+      title: 'Time Format',
+      description: 'What format to use for displaying time',
+      icon: Icons.hourglass_empty,
+      value: '',
+      values: patterns.map((pattern) {
+        return Option(
+          value: pattern.pattern,
+          text: pattern.format(date),
+        );
+      }),
+      onChanged: (v) {
+        settings.timeFormat = DateFormat(v!, locale);
+      },
+    );
   }
 }
